Rename to SDL device
