+++ source/Irrlicht/COpenGLCommon.h
@@ -42,18 +42,6 @@
 		#define GLX_GLXEXT_PROTOTYPES 1
 	#endif
 	#define NO_SDL_GLEXT
-	#include <SDL/SDL_video.h>
-	#include <SDL/SDL_opengl.h>
-	#include "glext.h"
-#elif defined(_IRR_COMPILE_WITH_SDL2_DEVICE_) && !defined(_IRR_COMPILE_WITH_X11_DEVICE_)
-	#if defined(_IRR_OPENGL_USE_EXTPOINTER_)
-		#define GL_GLEXT_LEGACY 1
-		#define GLX_GLXEXT_LEGACY 1
-	#else
-		#define GL_GLEXT_PROTOTYPES 1
-		#define GLX_GLXEXT_PROTOTYPES 1
-	#endif
-	#define NO_SDL_GLEXT
 	#include <SDL_video.h>
 	#include <SDL_opengl.h>
 	#include "glext.h"
